+++ src/config/topics.js
@@ -34,6 +34,7 @@ import StsMenu from '../menus/StsMenu';
 import { DV_KEY } from '../utils/constants';
 import DvMenu from '../menus/DirektverbindungenMenu/DvMenu';
 import DvListButton from './ch.sbb.direktverbindungen/DvListButton';
+import applPermalinkVisiblity from '../utils/applyPermalinkVisibility';
 
 // For backward compatibility
 export {
@@ -360,6 +361,7 @@ export const sts = {
   center: [915788.3813658276, 5909670.533831286],
   zoom: 8,
   constrainOnlyCenter: true,
+  overlaySide: 'left',
 };
 
 export const direktverbindungen = {
@@ -406,6 +408,7 @@ export const direktverbindungenIframe = {
   hideInLayerTree: true,
   menu: <DvMenu />,
   mapControls: <DvListButton />,
+  overlaySide: 'left',
 };
 
 const topics = {
@@ -431,6 +434,13 @@ const topics = {
   stelen: [netzkarteStelen],
   betriebsregionen: [betriebsregionen],
 };
+
+topics.wkp.forEach((topic) => {
+  applPermalinkVisiblity(
+    topic.layers,
+    (pathname) => pathname?.indexOf(`/${topic.key}`) !== -1,
+  );
+});
 
 export const getTopicConfig = (name) => {
   return topics[name];