+++ src/menus/DirektverbindungenMenu/DvMenu.js
@@ -1,16 +1,27 @@
 /* eslint-disable no-param-reassign */
 import React, { useEffect, useMemo } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { Divider, makeStyles } from '@material-ui/core';
+import { makeStyles, Divider } from '@material-ui/core';
 import DvFeatureInfo from '../../config/ch.sbb.direktverbindungen/DvFeatureInfo';
 import DvLayerSwitcher from './DvLayerSwitcher';
 import { setDisplayMenu } from '../../model/app/actions';
 import IframeMenu from '../IframeMenu';
+import useIsMobile from '../../utils/useIsMobile';
+import useHighlightLayer from '../../utils/useHighlightLayer';
+import DvFeatureInfoTitle from '../../config/ch.sbb.direktverbindungen/DvFeatureInfoTitle/DvFeatureInfoTitle';
 
 const useStyles = makeStyles(() => {
   return {
     wrapper: {
-      padding: '15px 10px',
+      padding: 15,
+    },
+    listHeader: {
+      padding: '5px 12px',
+      fontSize: 16,
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'space-between',
+      backgroundColor: '#f5f5f5',
     },
   };
 });
@@ -18,12 +29,10 @@ const useStyles = makeStyles(() => {
 function DvMenu() {
   const dispatch = useDispatch();
   const classes = useStyles();
-  const screenWidth = useSelector((state) => state.app.screenWidth);
   const featureInfo = useSelector((state) => state.app.featureInfo);
   const topic = useSelector((state) => state.app.activeTopic);
-  const isMobile = useMemo(() => {
-    return ['xs'].includes(screenWidth);
-  }, [screenWidth]);
+  const highlightLayer = useSelector((state) => state.map.highlightLayer);
+  const isMobile = useIsMobile();
 
   const urlSearch = useMemo(
     () => new URLSearchParams(window.location.search),
@@ -60,14 +69,18 @@ function DvMenu() {
     }
   }, [isMobile, topic, dispatch]);
 
+  // Hook to highlight map features
+  useHighlightLayer(featureInfo, highlightLayer);
+
   return (
     <IframeMenu
       hide={hideMenu}
       header={showSwitcher ? switcher : null}
+      title={<DvFeatureInfoTitle />}
       body={
         <>
           {showSwitcher && isMobile ? switcher : null}
-          {showSwitcher ? <Divider /> : null}
+          {showSwitcher && isMobile ? <Divider /> : null}
           <DvFeatureInfo filterByType />
         </>
       }