+++ src/utils/constants.js
@@ -46,7 +46,7 @@ export const energieleitungenColorMapping = {
 export const SWISS_EXTENT = [656409.5, 5740863.4, 1200512.3, 6077033.16];
 
 // UI
-export const OVERLAY_MIN_HEIGHT = 50;
+export const OVERLAY_MIN_HEIGHT = 55;
 
 // PDF export
 export const FORCE_EXPORT_PROPERTY = 'forceVisibilityOnPdf';