+++ src/utils/dvParseFeatures.js
@@ -3,21 +3,18 @@
  */
 const parseDvFeatures = (featuresArray) => {
   return featuresArray.map((feature) => {
-    const {
-      start_station_name: start,
-      end_station_name: end,
-      vias,
-    } = feature.getProperties();
+    const { vias } = feature.getProperties();
+    const parsedVias = Array.isArray(vias) ? vias : JSON.parse(vias);
 
     const switchVias = vias
-      ? (Array.isArray(vias) ? vias : JSON.parse(vias)).filter(
+      ? parsedVias.filter(
           (via) => via.via_type === 'switch' || via.via_type === 'visible',
         )
       : [];
     feature.set('vias', [
-      { station_name: start },
+      parsedVias[0],
       ...switchVias,
-      { station_name: end },
+      parsedVias[parsedVias.length - 1],
     ]);
     return feature;
   });